Moved the helper services to the same docker compose
The issue was with the communication between the server and the databases. Docker compose is fond of connecting to a host service. I guess it worked before as the helper services was on an external server; however, that is nto certain. Anyways, I moved the services to the same compose file and used the server_default network such that the reverse proxy could recognize it.