Added images and a for loop to display the skills section
The images are stores in a txt file called skills_links.txt . The backend reads the images from the txt file and forwards it to the front end. In the frontend, a for loop puts the images in the skills section neatly. Had to add padding to get the correct spacing between rows.